fix: 이자율 자료형 변경 및 2024-03-01까지의 변경사항 저장

+++ src/main/java/com/fpr/financialProduct/dto/FinancialProductOption.java
@@ -19,9 +19,9 @@ public class FinancialProductOption {
     private String saveTerm;
 
     @JsonProperty("intr_rate")
-    private int interestRate;
+    private double interestRate;
 
     @JsonProperty("intr_rate2")
-    private int interestRate2;
+    private double interestRate2;
 
 }
